# server/openapi.yaml
openapi: 3.0.3
info:
  title: Password Protector API
  version: 1.0.0
  description: |
    REST API for Password Protector — a zero-knowledge password manager with client-side AES-256-GCM encryption.

    Authentication endpoints are public; password and user endpoints require a valid JWT Bearer token obtained from the login or refresh flows.

    The server never sees plaintext passwords. Passwords are encrypted client-side before being sent to the API.
servers:
  - url: http://localhost:3000/api
    description: Local development
  - url: https://{domain}/api
    description: Deployed environment
    variables:
      domain:
        default: password-protector.shalev396.com
...
security: []
paths:
...
  /private/me:
    get:
      tags: [User]
      summary: Get current user profile
      description: Returns the authenticated user's profile information.
      operationId: getMe
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    put:
      tags: [User]
      summary: Update current user profile
      description: Updates the authenticated user's profile. Currently supports updating the user's name.
      operationId: updateMe
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMeRequestBody'
            example:
              name: Jane Smith
      responses:
        '200':
          description: Updated user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /private/delete:
    delete:
      tags: [User]
      summary: Delete user account
      description: Permanently deletes the authenticated user's account, all associated password entries, and the Cognito user record.
      operationId: deleteAccount
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Account deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteUserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  # -- Dev ---------------------------------------------------------------------
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Cognito-issued JWT ID token

  # -- Reusable Responses ------------------------------------------------------
  responses:
    BadRequest:
      description: Invalid or missing request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          example:
            message: 'Title and password are required'
    Unauthorized:
      description: Authentication required or credentials invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          example:
            message: Login failed
    NotFound:
      description: Requested resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          example:
            message: Password not found
...
    InternalError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          example:
            message: An unexpected error occurred

  # -- Schemas -----------------------------------------------------------------
  schemas:
...
    ApiErrorResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
...
    MeResponseData:
      type: object
      required: [id, cognitoSub, email, name, lastLoginAt, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        cognitoSub:
          type: string
        email:
          type: string
        name:
          type: string
          nullable: true
        lastLoginAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    MeResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: '#/components/schemas/MeResponseData'
    UpdateMeRequestBody:
      type: object
      properties:
        name:
          type: string
    DeleteUserResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [message]
          properties:
            message:
              type: string

    # -- Dev response data -----------------------------------------------------
...